requirements.txt
flask 
ultralytics
flask_cors
waitress
easyocr
matplotlib
asyncio
python-dotenv
Pillow